Move users repository outside internal
So external users can instantiate postgres repository
